Decouple Shopify from Store
