+++ BTCPayServer.Tests/BTCPayServer.Tests.csproj
@@ -53,4 +53,8 @@
     <ProjectReference Include="..\BTCPayServer\BTCPayServer.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Pages" />
+  </ItemGroup>
+
 </Project>
