+++ BTCPayServer/Models/StoreViewModels/IntegrationsViewModel.cs
@@ -10,6 +10,5 @@ namespace BTCPayServer.Models.StoreViewModels
 {
     public class IntegrationsViewModel
     {
-        public ShopifySettings Shopify { get; set; }
     }
 }
