+++ BTCPayServer/Shopify/ApiModels/CreateScriptResponse.cs
