+++ BTCPayServer/Shopify/ApiModels/CreateWebhookResponse.cs
