+++ BTCPayServer/Shopify/ApiModels/DataHolders/TransactionDataHolder.cs
