+++ BTCPayServer/Shopify/ApiModels/ShopifyTransaction.cs
