+++ BTCPayServer/Shopify/ApiModels/TransactionsCreateResp.cs
