+++ BTCPayServer/Shopify/ApiModels/TransactionsListResp.cs
