+++ BTCPayServer/Shopify/ShopifyApiException.cs
