+++ BTCPayServer/Views/Stores/Integrations.cshtml
@@ -4,7 +4,7 @@
     ViewData.SetActivePageAndTitle(StoreNavPages.Integrations, "Integrations");
 }
 
-<partial name="_StatusMessage" />
+<partial name="_StatusMessage"/>
 
 @if (!ViewContext.ModelState.IsValid)
 {
@@ -15,20 +15,30 @@
     </div>
 }
 
-<div class="row">
-    <div class="col-md-8">
-       <partial name="Integrations/Shopify"/>
+<ul class="list-group mb-3">
+    <vc:ui-extension-point location="store-integrations-list"></vc:ui-extension-point>
+</ul>
 
+<div class="row">
+    <div class="col-12">
         <h4 class="mb-3 mt-5">
             Other Integrations
         </h4>
         <p>
             Take a look at documentation for the list of other integrations we support and the directions on how to enable them:
             <ul>
-                <li><a href="https://docs.btcpayserver.org/WooCommerce/" target="_blank">WooCommerce</a></li>
-                <li><a href="https://docs.btcpayserver.org/Drupal/" target="_blank">Drupal</a></li>
-                <li><a href="https://docs.btcpayserver.org/Magento/" target="_blank">Magento</a></li>
-                <li><a href="https://docs.btcpayserver.org/PrestaShop/" target="_blank">PrestaShop</a></li>
+                <li>
+                    <a href="https://docs.btcpayserver.org/WooCommerce/" target="_blank">WooCommerce</a>
+                </li>
+                <li>
+                    <a href="https://docs.btcpayserver.org/Drupal/" target="_blank">Drupal</a>
+                </li>
+                <li>
+                    <a href="https://docs.btcpayserver.org/Magento/" target="_blank">Magento</a>
+                </li>
+                <li>
+                    <a href="https://docs.btcpayserver.org/PrestaShop/" target="_blank">PrestaShop</a>
+                </li>
             </ul>
         </p>
     </div>
